+++ core/src/main/java/org/web3j/protocol/core/Request.java
@@ -15,7 +15,6 @@ package org.web3j.protocol.core;
 import java.io.IOException;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.atomic.AtomicLong;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import io.reactivex.Flowable;
@@ -23,8 +22,6 @@ import io.reactivex.Flowable;
 import org.web3j.protocol.Web3jService;
 
 public class Request<S, T extends Response> {
-    private static AtomicLong nextId = new AtomicLong(0);
-
     private String jsonrpc = "2.0";
     private String method;
     private List<S> params;
@@ -41,7 +38,7 @@ public class Request<S, T extends Response> {
     public Request(String method, List<S> params, Web3jService web3jService, Class<T> type) {
         this.method = method;
         this.params = params;
-        this.id = nextId.getAndIncrement();
+        this.id = DefaultIdProvider.getNextId();
         this.web3jService = web3jService;
         this.responseType = type;
     }